+++ naga/src/proc/namer.rs
@@ -1,16 +1,15 @@
 use alloc::{
     borrow::Cow,
-    boxed::Box,
     format,
     string::{String, ToString},
     vec::Vec,
 };
-use core::hash::{Hash, Hasher};
 
-use hashbrown::HashSet;
-use once_cell::race::OnceBox;
-
-use crate::{arena::Handle, FastHashMap, FastHashSet};
+use crate::{
+    arena::Handle,
+    proc::{keyword_set::CaseInsensitiveKeywordSet, KeywordSet},
+    FastHashMap,
+};
 
 pub type EntryPointIndex = u16;
 const SEPARATOR: char = '_';
@@ -86,25 +85,13 @@ pub enum NameKey {
 
 /// This processor assigns names to all the things in a module
 /// that may need identifiers in a textual backend.
+#[derive(Default)]
 pub struct Namer {
     /// The last numeric suffix used for each base name. Zero means "no suffix".
     unique: FastHashMap<String, u32>,
-    keywords: &'static HashSet<&'static str>,
-    keywords_case_insensitive: FastHashSet<AsciiUniCase<&'static str>>,
+    keywords: &'static KeywordSet,
+    keywords_case_insensitive: &'static CaseInsensitiveKeywordSet,
     reserved_prefixes: Vec<&'static str>,
-}
-
-impl Default for Namer {
-    fn default() -> Self {
-        static DEFAULT_KEYWORDS: OnceBox<HashSet<&'static str>> = OnceBox::new();
-
-        Self {
-            unique: Default::default(),
-            keywords: DEFAULT_KEYWORDS.get_or_init(|| Box::new(HashSet::default())),
-            keywords_case_insensitive: Default::default(),
-            reserved_prefixes: Default::default(),
-        }
-    }
 }
 
 impl Namer {
@@ -191,13 +178,11 @@ impl Namer {
                 let mut suffixed = base.to_string();
                 if base.ends_with(char::is_numeric)
                     || self.keywords.contains(base.as_ref())
-                    || self
-                        .keywords_case_insensitive
-                        .contains(&AsciiUniCase(base.as_ref()))
+                    || self.keywords_case_insensitive.contains(base.as_ref())
                 {
                     suffixed.push(SEPARATOR);
                 }
-                debug_assert!(!self.keywords.contains::<str>(&suffixed));
+                debug_assert!(!self.keywords.contains(&suffixed));
                 // `self.unique` wants to own its keys. This allocates only if we haven't
                 // already done so earlier.
                 self.unique.insert(base.into_owned(), 0);
@@ -228,8 +213,8 @@ impl Namer {
     pub fn reset(
         &mut self,
         module: &crate::Module,
-        reserved_keywords: &'static HashSet<&'static str>,
-        reserved_keywords_case_insensitive: &[&'static str],
+        reserved_keywords: &'static KeywordSet,
+        reserved_keywords_case_insensitive: &'static CaseInsensitiveKeywordSet,
         reserved_prefixes: &[&'static str],
         output: &mut FastHashMap<NameKey, String>,
     ) {
@@ -238,16 +223,7 @@ impl Namer {
 
         self.unique.clear();
         self.keywords = reserved_keywords;
-
-        debug_assert!(reserved_keywords_case_insensitive
-            .iter()
-            .all(|s| s.is_ascii()));
-        self.keywords_case_insensitive.clear();
-        self.keywords_case_insensitive.extend(
-            reserved_keywords_case_insensitive
-                .iter()
-                .map(|string| AsciiUniCase(*string)),
-        );
+        self.keywords_case_insensitive = reserved_keywords_case_insensitive;
 
         // Choose fallback names for anonymous entry point return types.
         let mut entrypoint_type_fallbacks = FastHashMap::default();
@@ -380,33 +356,6 @@ impl Namer {
             };
             let name = self.call(label);
             output.insert(NameKey::Constant(handle), name);
-        }
-    }
-}
-
-/// A string wrapper type with an ascii case insensitive Eq and Hash impl
-struct AsciiUniCase<S: AsRef<str> + ?Sized>(S);
-
-impl<S: AsRef<str>> PartialEq<Self> for AsciiUniCase<S> {
-    #[inline]
-    fn eq(&self, other: &Self) -> bool {
-        self.0.as_ref().eq_ignore_ascii_case(other.0.as_ref())
-    }
-}
-
-impl<S: AsRef<str>> Eq for AsciiUniCase<S> {}
-
-impl<S: AsRef<str>> Hash for AsciiUniCase<S> {
-    #[inline]
-    fn hash<H: Hasher>(&self, hasher: &mut H) {
-        for byte in self
-            .0
-            .as_ref()
-            .as_bytes()
-            .iter()
-            .map(|b| b.to_ascii_lowercase())
-        {
-            hasher.write_u8(byte);
         }
     }
 }